+++ dantro/data_ops/_base_ops.py
@@ -1,6 +1,8 @@
-"""Implements operations that need to be importable from other modules"""
+"""Implements operations that need to be importable from other modules or that
+are so basic that they apply to a wide range of applications."""
 
 import operator
+from typing import Callable
 
 # -----------------------------------------------------------------------------
 
@@ -25,3 +27,27 @@ BOOLEAN_OPERATORS = {
 } # End of boolean operator definitions
 """Boolean binary operators"""
 # fmt: on
+
+
+# -----------------------------------------------------------------------------
+
+
+def _make_passthrough(func: Callable) -> Callable:
+    """Wraps a callable such that it returns its first positional argument.
+
+    This is meant for functions that operate on an object (conventionally the
+    first argument) and do not have a return value.
+    By constructing a callable using this function, it can be made compatible
+    with the data transformation framework.
+
+    .. code-block:: python
+
+        f = setattr                 # f has no return value
+        g = _make_passtrough(f)     # g will return the first argument
+    """
+
+    def wrapped(d, *args, **kwargs):
+        func(d, *args, **kwargs)
+        return d
+
+    return wrapped